ref(api): Move discard and some deletion code from project index to mixin
